Backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
supabase>=2.3.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.7
passlib[bcrypt]>=1.7.4
alembic>=1.13.1
python-dotenv>=1.0.0
gunicorn>=21.2.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.19
pydantic>=2.6.0
